+++ app/javascript/application.js
@@ -3,7 +3,6 @@ import "@hotwired/turbo-rails"
 import "./controllers"
 import { initializeButtonsAndModal } from "./controllers/modal";
 import "./controllers/search";
-import "./controllers/search_conditions";
 
 document.addEventListener('turbo:load', initializeButtonsAndModal);
 document.addEventListener('turbo:render', initializeButtonsAndModal);
