+++ app/views/spotify/results.html.erb
@@ -1,11 +1,6 @@
 <!-- 📊 検索結果エリア -->
 <h3 class="text-lg font-medium text-white mb-4 text-center">🎧 検索結果</h3>
 
-<!-- 検索フォーム -->
-<div class="p-6 rounded-lg shadow-md mb-8">
-  <%= render 'spotify/search' %>
-</div>
-
 <% if @tracks.present? %>
   <div class="p-6 bg-gray-800 rounded-lg shadow-md">
     <ul class="space-y-4">
@@ -36,7 +31,44 @@
   <p class="text-white text-center">🔍 検索結果が見つかりませんでした。</p>
 <% end %>
 
-<!-- 🔄 検索フォームに戻るリンク -->
-<div class="text-center mt-6">
-  <%= link_to "📓 日記作成ページに戻る", new_journal_path, class: "px-4 py-2 bg-gray-500 text-white rounded hover:bg-gray-600" %>
+<!-- 検索フォーム -->
+<div class="p-6 rounded-lg shadow-md mb-8">
+  <div class="p-6 rounded-lg bg-customblue ">
+    <%= form_with url: spotify_search_path, method: :get, local: false, remote: true, id: "spotify-search-form" do |f| %>
+      <h2 class="text-xl font-bold mb-6 text-white text-center">🎵 再検索フォームはこちら</h2>
+      <!-- 🔍 検索タイプ -->
+      <div class="mb-4">
+        <%= f.label :initial_search_type, "🔍 検索タイプ", class: "block text-sm font-medium text-white mb-2" %>
+        <%= f.select :initial_search_type, 
+              [["曲名", "track"], ["アーティスト名", "artist"], ["キーワード", "keyword"]],
+              { prompt: "検索タイプを選択" },
+              { id: "initial-search-type",
+                class: "block w-full px-4 py-2 border border-gray-400 rounded-md text-white bg-gray-700 focus:ring-blue-500 focus:border-blue-500" } %>
+      </div>
+
+      <!-- 📝 検索キーワード -->
+      <div class="mb-6">
+        <%= f.label :initial_query, "📝 検索キーワード", class: "block text-sm font-medium text-white mb-2" %>
+        <%= f.text_field :initial_query, placeholder: "キーワードを入力", id: "initial-query", 
+              class: "block w-full px-4 py-2 border border-gray-400 rounded-md text-white bg-gray-700 placeholder-gray-400 focus:ring-blue-500 focus:border-blue-500" %>
+      </div>
+
+      <!-- ➕ 検索条件の追加・削除 -->
+      <div id="search-conditions" class="space-y-4 mb-4"></div>
+
+      <div class="flex gap-4 mb-6">
+        <button type="button" id="add-condition-btn" class="flex-1 px-4 py-2 bg-green-500 text-white font-medium rounded hover:bg-green-600 text-center">
+          ➕ 検索条件を追加
+        </button>
+        <button type="button" id="remove-condition-btn" class="flex-1 px-4 py-2 bg-red-500 text-white font-medium rounded hover:bg-red-600 text-center">
+          ➖ 検索条件を削除
+        </button>
+      </div>
+
+      <!-- 🔎 検索ボタン -->
+      <div class="text-center mb-4">
+        <%= f.submit "🔎 検索", class: "w-full px-4 py-2 bg-blue-500 text-white font-medium rounded hover:bg-blue-600 text-center" %>
+      </div>
+    <% end %>
+  </div>
 </div>
